refactor(codegen): extract checked arithmetic lowering
